requirements.txt
python-telegram-bot==21.7
openai==1.58.0
pyyaml==6.0.2
python-dotenv==1.0.1
